+++ app/Http/Controllers/Admin/SolarController.php
@@ -1,21 +1,16 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
-
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-use App\Traits\Solar;
-use Mail;
+use App\Traits\Solar; use Mail;
 use App\Traits\SolarPosting;
 use App\Models\SaleRecord;
-use App\Models\Project;
-use DB;
-use App\User;
-use App\Http\Requests\SolarRequest;
-use App\Exports\ExportSolar;
+use App\Models\Project;use DB;
+use App\User;use App\Http\Requests\SolarRequest;
+use App\Exports\ExportSolar; 
 use Maatwebsite\Excel\Facades\Excel;
-use App\Models\Client;
-use Auth;
+use App\Models\Client;use Auth;
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 
@@ -23,7 +18,7 @@ use Spatie\Permission\Models\Permission;
 class SolarController extends Controller
 {
 
-    use Solar, SolarPosting;
+    use Solar,SolarPosting;
     /**
      * Create a new controller instance.
      *
@@ -42,34 +37,30 @@ class SolarController extends Controller
      */
     public function index(Request $request)
     {
-        $user = auth()->user();
-        $raw_result = DB::select(DB::raw("SELECT * FROM model_has_roles WHERE model_id = '$user->id'"));
-        $roleid = $raw_result[0]->role_id;
-
-        if (!in_array($roleid, array(14, 15, 1, 18, 3))) {
-            return redirect()->route('solar.create')->with('error', "Access Denied");
-        }
-
-        if ($request->export) {
-            return Excel::download(new ExportSolar(@$request->start_date, @$request->end_date, @$request->search, @$request->client_id, Auth::user()->hasRole('SolarClient'), Auth::user()->id), 'ExportSolar.xlsx');
+		$user = auth()->user(); 
+        if($request->export){
+            return Excel::download(new ExportSolar(@$request->start_date,@$request->end_date,
+			@$request->search,@$request->client_id,@$request->project_id,Auth::user()->id), 'ExportSolar.xlsx');
         }
         $states = DB::table("electric_provider")->get();
-        $clients = Client::where('campaign_id', 2)->get();
-        $projects = Project::whereIn('client_id', $clients->pluck('id'))->get();
-        $this->authorize('solars.index');
+        $clients = Client::where('campaign_id',2)->get();
+        $projects = Project::whereIn('client_id',$clients->pluck('id'))->get();
+        $this->authorize('solars.index');  
         $search = @$request->search;
         $start_date = @$request->start_date;
         $end_date = @$request->end_date;
+        $solars  =   SaleRecord::with('user','client')->where('campaign_id',"2");
+        if(Auth::user()->hasRole('SolarClient')){			
+            $project_codes = User::with('projects')->where('id',Auth::user()->id)->get()->pluck('projects')->flatten()->pluck('project_code');
+            $solars  =   $solars->whereIn('project_code',$project_codes);
+        } 
+        if(($request->search) || ($request->start_date) || ($request->end_date) || ($request->client_id) || ($request->project_id))
+        {
+            $solars = $solars->search($request->search,@$request->start_date,@$request->end_date,@$request->client_id,@$request->project_id);
+        } 
+        $solars = $solars->orderby('id','DESC' )->paginate(100);
+        return view('admin.solar.index',compact('solars','clients','projects','states'));
 
-        if (Auth::user()->hasRole('SolarClient')) {
-            $project_codes = User::with('projects')->where('id', Auth::user()->id)->get()->pluck('projects')->flatten()->pluck('project_code');
-            $solars  =   SaleRecord::with('user', 'client')->where('campaign_id', "2")->whereIn('project_code', $project_codes);
-        } else {
-            $solars  =   SaleRecord::with('user', 'client')->where('campaign_id', "2");
-        }
-        $solars = $solars->search($request->search, @$request->start_date, @$request->end_date, @$request->client_id, @$request->project_id)
-            ->orderby('id', 'DESC')->paginate(8);
-        return view('admin.solar.index', compact('solars', 'clients', 'projects', 'states'));
     }
 
     /**
@@ -80,10 +71,10 @@ class SolarController extends Controller
     public function create()
     {
         $states = DB::table("electric_provider")->groupBy('state')->get();
-        $this->authorize('solars.create');
-        $clients = Client::where('campaign_id', 2)->pluck('id');
-        $projects = Project::whereIn('client_id', $clients)->get();
-        return view('admin.solar.create', compact('clients', 'projects', 'states'));
+        $this->authorize('solars.create'); 
+        $clients = Client::where('campaign_id',2)->pluck('id');
+        $projects = Project::whereIn('client_id',$clients)->get();
+        return view('admin.solar.create',compact('clients','projects','states'));
     }
 
     /**
@@ -93,29 +84,43 @@ class SolarController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(SolarRequest $request)
-    {
+    { 
+		//$this->authorize('solars.store');  
+        if($request->record_id<=0){
+            return redirect()->route('solars.create')->with('error',"RecordID Required");
+        }
+        $request->validate([
+            'first_name'=>"required",
+            'last_name'=>"required",
+            'phone'=>"required",
+            'clients'=>"required",
+            'record_id'=>"required",
+        ]);
         DB::connection('mysql')->beginTransaction();
-        try {
-            $check = DB::table('sale_records')->where('phone', $request->phone)
-                ->whereDate('created_at', ">=", date('Y-m-d', strtotime('-90 days')))->first();
-            $checkOld = DB::table('phone_numbers')->where('phone', $request->phone)->first();
-            if ($check || $checkOld) {
-                return redirect()->back()->with('error', 'Phone no already used');
-            }
+        try{ 
+            
+            $check =    DB::table('sale_records')->where('phone',$request->phone)->whereDate('created_at',"<=",date('Y-m-d',strtotime('-90 days')) )->first();                    
+            $check1 =   DB::table('sale_records')->where('record_id',$request->record_id)->first(); 
+            $checkOld = DB::table('phone_numbers')->where('phone',$request->phone)->first();
+            if($check || $checkOld || $check1){ 
+                return redirect()->back()->with('error','Phone no already used');
+            } 
             $data = $request->all();
-            $res = $this->InsertSaleRecord($data);
-            //$this->send_mial($res,"solar@excelcg.com");
-            $res = $this->postingUrl($request->clients, $data, $res);
+            $res = $this->InsertSaleRecord($data); 
+            $res = $this->postingUrl($request->clients,$data,$res);  
             DB::commit();
-
-            return redirect()->route('solars.create')->with('success', "Post Successfully");
-        } catch (\Exception $e) {
-            DB::connection('mysql')->rollback();
-            return redirect()->route('solars.create')->with('error', $e->getMessage());
+			
+            return redirect()->route('solars.create')->with('success',"Post Successfully");
+        }catch(\Exception $e){
+            DB::connection('mysql')->rollback(); 
+			if(Auth::user()->hasRole('Super Admin')){
+              //return redirect()->route('solars.create')->with('error',$e->getMessage());
+			}
+            return redirect()->route('solars.create')->with('error',"Internal Server Error");
         }
     }
-
-
+	
+	
 
     /**
      * Display the specified resource.
@@ -125,8 +130,9 @@ class SolarController extends Controller
      */
     public function show($id)
     {
-        $data = SaleRecord::where('id', $id)->first();
-        return view('admin.solar.view', compact('data'));
+		$this->authorize('solars.show');  
+        $data = SaleRecord::where('id',$id)->first();
+        return view('admin.solar.view',compact('data'));
     }
 
     /**
@@ -138,6 +144,7 @@ class SolarController extends Controller
     public function edit($id)
     {
         //
+		$this->authorize('solars.edit');  
     }
 
     /**
@@ -150,6 +157,7 @@ class SolarController extends Controller
     public function update(Request $request, $id)
     {
         //
+		$this->authorize('solars.update'); 
     }
 
     /**
@@ -160,7 +168,18 @@ class SolarController extends Controller
      */
     public function destroy($id)
     {
-        SaleRecord::where('id', $id)->delete();
-        return redirect()->back()->with("success", "Disabled successfully");
+		$this->authorize('solars.destroy'); 
+        SaleRecord::where('id',$id)->delete();
+        return redirect()->back()->with("success","Disabled successfully");
     }
+    public function devsolar($id)
+    {    
+        $states = DB::table("electric_provider")->groupBy('state')->get();
+        $this->authorize('solars.create'); 
+        $clients = Client::where('campaign_id',2)->pluck('id');
+        $projects = Project::whereIn('client_id',$clients)->get();
+        return view('admin.solar.devcreate',compact('clients','projects','states'));
+    }
+    
+    
 }
